Otimizacao das pastas e alteracao da NavBar
