+++ index.html
@@ -1,4 +1,3 @@
-
 <!doctype html>
 <html lang="en">
   <head>
