+++ src/components/Navbar/Navbar.css
@@ -1,0 +1,65 @@
+.main-content {
+    padding-top: 70px;
+}
+
+.navbar {
+    background-color: #003366 !important; 
+    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
+    text-align: center;
+  }
+
+  .navbar-nav {
+  justify-content: center;
+  width: 100%;
+}
+  
+  .fixed-navbar {
+    position: fixed;
+    top: 0;
+    z-index: 1020;
+    width: 100%;
+  }  
+
+  .navbar-brand {
+    font-size: 1.75rem;
+    font-weight: bold;
+    color: #ffffff !important;
+  }
+  
+  .nav-link {
+    color: #ffffff !important;
+    margin-left: 1rem;
+    font-weight: 500;
+    position: relative;
+    transition: color 0.3s ease;
+  }
+  
+  .nav-link::after {
+    content: '';
+    position: absolute;
+    bottom: -5px;
+    left: 0;
+    width: 0%;
+    height: 2px;
+    background-color: #7c9bff;
+    transition: 0.3s ease-in-out;
+  }
+  
+  .nav-link:hover {
+    color: #7c9bff !important;
+  }
+  
+  .nav-link:hover::after {
+    width: 100%;
+  }
+  
+  @media (max-width: 768px) {
+    .navbar {
+      flex-direction: column;
+      align-items: flex-start;
+    }
+  
+    .nav-link {
+      margin: 0.5rem 0;
+    }
+  }
