+++ src/components/Tabelabancos/TabelaBancos.css
@@ -1,0 +1,6 @@
+.main-content {
+    padding-top: 80px; /* Ajuste conforme a altura real da sua navbar */
+    padding-left: 20px;
+    padding-right: 20px;
+  }
+
