+++ src/css.module.css
@@ -7,6 +7,16 @@
     margin-top: 2px;
     color: var(--bs-secondary);
 }
+body {
+    padding-top: 70px;
+  }
+  
+  @media (max-width: 768px) {
+    body {
+      padding-top: 80px; /* navbar pode ficar maior no mobile */
+    }
+  }
+  
 :root {
     font-family: Arial, Helvetica, sans-serif;
     display: flex;
@@ -28,7 +38,7 @@ table{
     width: min-content;
     & td{
         
-        width: 25%; /* Ajuste a largura conforme necessário */
+        width: 25%;
         
     }
